% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{find_best_fit}
\alias{find_best_fit}
\title{Find the most fit model by grid search of number of models}
\usage{
find_best_fit(data, f, criteria = "bic")
}
\arguments{
\item{data}{The dataframe to pass in}

\item{f}{the formula for the LDA modeling}

\item{criteria}{The criteria to choose the best model, default to BIC}
}
\value{
The model object of the most fit model according to the criteria
}
\description{
Find the most fit model by grid search of number of models
}
\examples{
# Define a formula for the LDA modeling
f = with(data, cbind(tax, religion, free_election, state_aid, civil_rights, women)~1)
find_best_fit(data, f, criteria = 'bic') # This will yield the plot


}
