% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{profile_plot}
\alias{profile_plot}
\title{Function to produce a profile plot for a LDA model object}
\usage{
profile_plot(data, model = NULL, num_var, f)
}
\arguments{
\item{data}{The dataframe to pass in}

\item{model}{the most fit model produced by the function find_best_fit}

\item{num_var}{The number of the variables put into the LCA modeling}

\item{f}{the formula for the LDA modeling}
}
\value{
The ggplot objects
}
\description{
Function to produce a profile plot for a LDA model object
}
\examples{
# Define a formula for the LDA modeling
f = with(data, cbind(tax, religion, free_election, state_aid, civil_rights, women)~1)
profile_plot(data, num_var, f) # This will yield the plot






}
