% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{stacked_bar_plot}
\alias{stacked_bar_plot}
\title{Visualization of each latent class in a stacked bar plot fashion}
\usage{
stacked_bar_plot(model)
}
\arguments{
\item{model}{This parameter expects to be the fitted latent class model object}
}
\value{
The stacked bar plot object; the plot will be printed
}
\description{
Visualization of each latent class in a stacked bar plot fashion
}
