% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_func.R
\name{test_func}
\alias{test_func}
\title{This is a test func}
\usage{
test_func(n)
}
\arguments{
\item{n}{A number}
}
\value{
hellow world plus n
}
\description{
This is a test func
}
\examples{
a = test_func(3)
print(a)
}
