% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{find_best_fit}
\alias{find_best_fit}
\title{Find the most fit model by grid search of number of models}
\usage{
find_best_fit(data, form, criterion = "bic")
}
\arguments{
\item{data}{The dataframe to pass in}

\item{form}{the formula for the LDA modeling}

\item{criterion}{The criterion to choose the best model, default to BIC.
Possible options include "aic",}
}
\value{
The model object of the most fit model according to the criterion
}
\description{
Find the most fit model by grid search of number of models
}
\examples{
# Define a formula for the LDA modeling
f = with(data, cbind(tax, religion, free_election, state_aid, civil_rights, women)~1)
find_best_fit(data, form = f, criterion = 'bic') # This will yield the plot


}
