% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{stacked_bar_by_class}
\alias{stacked_bar_by_class}
\title{Visualization of each latent class in a stacked bar plot fashion (facet by classes)}
\usage{
stacked_bar_by_class(model, color_palette = "Greys")
}
\arguments{
\item{model}{This parameter expects to be the fitted latent class model object}

\item{color_palette}{The color palette for the plot; default at "Greys," you could change to any
palette from brewer.pal from RColorBrewer package}
}
\value{
The stacked bar plot object; the plot will be printed
}
\description{
Visualization of each latent class in a stacked bar plot fashion (facet by classes)
}
\examples{
# Find out the best model according to the BIC criterion
best_model = find_best_fit(data, form, "bic")
# Pass the model object into the stacked_bar_plot function
stacked_bar_plot(best_model) # The function will print the plot and return the plot object

}
