% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_plot.R
\name{stacked_bar_by_item}
\alias{stacked_bar_by_item}
\title{Visualization of each latent class in a stacked bar plot fashion (facet by item variables)}
\usage{
stacked_bar_by_item(model, color_palette = "Greys")
}
\arguments{
\item{model}{This parameter expects to be the fitted latent class model object}

\item{color_palette}{The color palette for the plot; default at "Greys," you could change to any
palette from brewer.pal from RColorBrewer package}
}
\value{
The stacked bar plot object; the plot object will be printed and returned
}
\description{
Visualization of each latent class in a stacked bar plot fashion (facet by item variables)
}
